refactor(Classy) add select and dispatch types
