+++ src/resources/packages/classy/store/reducer.ts
@@ -1,5 +1,5 @@
 import { Action } from '@tec/common/classy/types/Actions';
-import { StoreState } from '../types/StoreState';
+import { StoreState } from '../types/Store';
 
 export const reducer = ( state: StoreState, action: Action ) => {
 	return state;
