--- src/resources/packages/classy/types/StoreState.d.ts
@@ -1,3 +1,0 @@
-export type StoreState = {
-	areTicketsSupported?: boolean;
-};
